+++ src/pages/index.tsx
@@ -1,20 +1,7 @@
 import { GetServerSidePropsContext } from 'next';
 import Head from 'next/head';
-import { useEffect } from 'react';
 
 export default function Home() {
-  const handleLogin = () => {
-    if (window.Kakao.isInitialized()) {
-      window.Kakao.Auth.authorize({
-        redirectUri: process.env.NEXT_PUBLIC_AUTH_REDIRECT_URL,
-      });
-    }
-  };
-
-  useEffect(() => {
-    window.Kakao.init(process.env.NEXT_PUBLIC_KAKAO_JAVASCRIPT_KEY);
-  }, []);
-
   return (
     <>
       <Head>
@@ -25,7 +12,6 @@ export default function Home() {
       </Head>
       <main>
         <h1>포텐파이낸스</h1>
-        <button onClick={handleLogin}>카카오 로그인</button>
       </main>
       <footer></footer>
     </>
